Fixed major issues to build on Netbeans 18
Now packages 2 JAR files: one with no dependences, and with the script to run that corrected, and one with dependences, that will  need to be built for each platform.

The first should in principle work on any platform on which Java is installed and a JavaFX library is available -- tested with v21 on ARM M1 Mac only so ofar.