Small updates to prevent bugs breaking saved file
Basis for new release. The hiding an individual or group in a PCA chart for now is partly disabled. Hiding a group works onscreen but does not save. Hiding an individual is turned off.

These features are not correctly implemented; they work onscreen but saving hidden items in a file -- even if they are unhidden before saving -- throws an exception when loading the file:

java.io.WriteAbortedException: writing aborted;
java.io.NotSerializable
Exception:<br>javafx.scene.chart.XYChart$Series

This is odd as the message relates to writing when the file is actually being read,