Handling hiding partially correct and file save cleaned up
Hidden individuals and groups now work to a limited extent; they can be hidden and the resulting view can be saved. However, unhiding does not always work.

Now when a project is saved as a new name, when you save again, the new name is offered as the name to use. To be more consistent with other apps, if you save as the same name, it should be easier than renaming.