checks that all lines (rows) of .phe and .fam files same length
If any line (row) of a .phe or .fam file has a different number of fields (columns) than the first line of that file, an error is reported showing the line number and creating a new project stops.

You can create a project with one or both file types; a project with only a .fam doesn’t make progress.