problem of hiding and reshowing across saves fixed
Before, if a group was hidden and even if re-shown, the saved file would not open.

The reason for this: the state was maintained in class PCAGraph in a hashmap, which it seems does not serialise cleanly. The simple solution would be to make the hashmap transient and recreate it each time except this would break opening older files so I made a new copy that is transient and don’t use the old one. This requires repopulating the hashmap after save-quit-load.