HashMap that doesn't serialise made transient
HashMap that was not saving and reloading correctly eliminated from save; project test file updated in case this is an issue (though the old one loads OK; this could jist be luck as deleting a class field could break serialise-reload and making a field `transient` has the effect of making it disappear from serialisation (as far as I can tell).

This fixed the problem of not being able to save hidden status of groups and recover this as saved.